Break up main.go and refactor inbound message handling
