+++ main.go
@@ -2,19 +2,12 @@ package main
 
 import (
 	"context"
-	"encoding/json"
 	"flag"
-	"os"
-	"os/signal"
-	"syscall"
-	"time"
 
 	"github.com/bwmarrin/discordgo"
-	"github.com/bytebot-chat/gateway-discord/model"
 	"github.com/go-redis/redis/v8"
 	"github.com/rs/zerolog"
 	"github.com/rs/zerolog/log"
-	uuid "github.com/satori/go.uuid"
 )
 
 // Variables used for command line parameters
@@ -30,6 +23,7 @@ var (
 	outbound  = flag.String("outbound", *id, "Pubsub to subscribe to for sending outbound messages. Defaults to being equivalent to `id`")
 )
 
+// init is run before main() and is used to initialize the logger
 func init() {
 	zerolog.TimeFieldFormat = zerolog.TimeFormatUnix
 
@@ -38,6 +32,8 @@ func init() {
 }
 
 func main() {
+
+	// Log our input parameters
 	log.Info().
 		Str("func", "main").
 		Str("id", *id).
@@ -46,165 +42,24 @@ func main() {
 		Str("redis", *redisAddr).
 		Msg("Starting Discord gateway")
 
-	rdb = rdbConnect(*redisAddr, *redisPass, 1) // Connect to redis
-	ctx = context.Background()                  // Redis context
-
-	// Create a new Discord session using the provided bot token.
-	dg, err := discordgo.New("Bot " + Token)
+	// Create a new Discordgo session using the provided bot token.
+	dgo, err := discordgo.New("Bot " + Token)
 	if err != nil {
 		log.Err(err).
 			Str("func", "main").
-			Msg("Unable to continue without connection. Exiting!")
-		os.Exit(1)
-	}
-
-	dg.AddHandler(func(s *discordgo.Session, r *discordgo.Ready) {
-		log.Info().
-			Str("func", "main").
-			Str("id", r.User.ID).
-			Str("username", r.User.Username).
-			Str("discriminator", r.User.Discriminator).
-			Msg("Discord gateway connected")
-	})
-	// Cleanly close down the Discord session.
-	defer dg.Close()
-	dg.AddHandler(messageCreate)                                                      // Register the messageCreate func as a callback for MessageCreate events.
-	dg.Identify.Intents = discordgo.MakeIntent(discordgo.IntentsAllWithoutPrivileged) // Listen for all low-privileged intents
-
-	err = dg.Open() // Open the websocket and begin listening.
-	if err != nil {
-		log.Err(err).
-			Str("func", "main").
-			Msg("Unable to open channel for reading messages")
-		os.Exit(1) // Exit if we can't open the channel
-	}
-
-	go handleOutbound(*outbound, rdb, dg) // Start listening for outbound messages
-
-	// Wait here until CTRL-C or other term signal is received.
-	log.Info().
-		Str("func", "main").
-		Msg("Bot is now running. Press CTRL-C to exit.")
-	sc := make(chan os.Signal, 1)                                             // Create a channel to listen for signals
-	signal.Notify(sc, syscall.SIGINT, syscall.SIGTERM, os.Interrupt, os.Kill) // Listen for signals
-	<-sc
-}
-
-// This function will be called (due to AddHandler above) every time a new
-// message is created on any channel that the authenticated bot has access to.
-func messageCreate(s *discordgo.Session, m *discordgo.MessageCreate) {
-
-	// Ignore all messages created by the bot itself
-	// This isn't required in this specific example but it's a good practice.
-	if m.Author.ID == s.State.User.ID {
+			Msg("Error creating Discord session")
 		return
 	}
 
-	str, _ := json.Marshal(m)         // Convert the message to JSON so that we can send it to Redis
-	msg := &model.Message{}           // Create a new message
-	err := msg.Unmarshal([]byte(str)) // Unmarshal the JSON into the message
-	if err != nil {
-		log.Err(err).
-			Str("func", "messageCreate").
-			Str("id", msg.ID).
-			Msg("Unable to unmarshal message")
-		return
-	}
+	r := redisConnect(*redisAddr, *redisPass, 0) // Connect to Redis
 
-	msg.Metadata.ID = uuid.Must(uuid.NewV4(), *new(error)) // Generate a UUID for the message
-	msg.Metadata.Source = *id                              // Set the source to the ID of this gateway
-	stringMsg, _ := json.Marshal(msg)                      // Convert the message to JSON so that we can send it to Redis
-	res := rdb.Publish(ctx, *inbound, stringMsg)           // Publish the message to Redis
-	if res.Err() != nil {
-		log.Err(res.Err()).
-			Str("func", "messageCreate").
-			Str("id", msg.ID).
-			Msg("Unable to publish message to Redis")
-		return
-	}
+	// Register the messageCreate func as a callback for MessageCreate events.
+	dgo.AddHandler(messageCreate)
 
-	log.Debug().
-		Str("func", "messageCreate").
-		Str("id", msg.Metadata.ID.String()).
-		Str("source", msg.Metadata.Source).
-		Str("dest", msg.Metadata.Dest).
-		Str("topic", *inbound).
-		Msg("Published message to Redis")
-}
+	// Open a connection to Redis pubsub and subscribe to the inbound topic
+	go handleInbound(*inbound, r, dgo)
 
-// handleOutbound handles outbound messages from Redis destined for Discord
-// It will subscribe to the given topic and send any messages it receives to Discord
-// Unforunately it uses a slightly different message format than the inbound messages
-// See model.MessageSend for more information
-func handleOutbound(sub string, rdb *redis.Client, s *discordgo.Session) {
-	log.Info().Str("func", "handleOutbound").Msg("Listening for outbound messages on topic '" + sub + "'")
-	ctx := context.Background()      // Redis context
-	topic := rdb.Subscribe(ctx, sub) // Subscribe to the given topic
-	channel := topic.Channel()       // Setup a Go channel to capture messages from Redis
+	// Hold the program open until CTRL-C is pressed
+	log.Info().Str("func", "main").Msg("Gateway is now running. Press CTRL-C to exit.")
 
-	for msg := range channel { // Loop through the messages as they come in
-		log.Debug().Str("func", "handleOutbound").
-			Str("topic", sub).
-			Str("payload", msg.Payload).
-			Send()
-
-		m := &model.MessageSend{}               // Create a new message
-		err := m.Unmarshal([]byte(msg.Payload)) // Unpack the message from Redis into Payload field
-		if err != nil {
-			log.Err(err).Str("func", "handleOutbound").Msg("Unable to unmarshal message") // TODO: Fix this log line
-		}
-
-		log.Debug().
-			Str("func", "handleOutbound").
-			Str("message_id", m.Metadata.ID.String()).
-			Str("message_source", m.Metadata.Source).
-			Str("message_destination", m.Metadata.Dest).
-			Bool("will_send", m.Metadata.Dest == *id).
-			Str("channel", m.ChannelID).
-			Str("content", m.Content).
-			Send()
-
-		if m.Metadata.Dest == *id { // Check if the message is for this gateway, if not, ignore it
-			_, err := s.ChannelMessageSend(m.ChannelID, m.Content) // Send the message to Discord
-			if err != nil {
-				log.Err(err).
-					Str("func", "handleOutbound").
-					Msg("Unable to send message to Discord")
-			}
-		}
-	}
-}
-
-// rdbConnect connects to Redis and returns a client
-func rdbConnect(addr, pass string, db int) *redis.Client {
-	log.Debug().Str("func", "rdbConnect").
-		Str("addr", addr).
-		Int("db", db).
-		Msg("Connecting to Redis")
-
-	ctx := context.Background() // Redis context
-	rdb := redis.NewClient(&redis.Options{
-		Addr:     addr, // Redis address
-		Password: pass, // Redis password
-		DB:       0,    // use default DB
-	})
-
-	err := rdb.Ping(ctx).Err() // Ping Redis to make sure we can connect
-	if err != nil {
-		// if we can't connect, try again in 3 seconds then succeed or give up
-		// yes this is lazy
-		log.Err(err).
-			Str("func", "rdbConnect").
-			Msg("Unable to connect to Redis, trying again in 3 seconds")
-		time.Sleep(3 * time.Second)
-		err := rdb.Ping(ctx).Err()
-		if err != nil {
-			log.Err(err).
-				Str("func", "rdbConnect").
-				Msg("Unable to connect to Redis, exiting")
-			os.Exit(1)
-		}
-	}
-
-	return rdb
 }
